requirements.txt
langchain==0.3.13
langchain-community==0.3.13
langchain-ollama==0.3.0
langchain-chroma==0.1.4
chromadb==0.5.23
pymupdf4llm==0.0.17
pandas==2.2.3
pydantic>=2.10.6
streamlit==1.41.1
python-dotenv>=1.0.1
crewai>=1.14.0
yfinance>=1.3.0
